Transcription
Write a simple program taking as input a DNA sequence stored into a file. The program must transcribe the input to RNA.
The result is displayed on the standard output. For instance, given a file with the following DNA content:
> Unknown
ACTGTTGTTCGGTGATCATCAGTTGTACAACGTCCTAACAACATCACATGCAATGCTTATGATATTCTTC
Your program would display the following information on the output:
ACUGUUGUUCGGUGAUCAUCAGUUGUACAACGUCCUAACAACAUCACAUGCAAUGCUUAUGAUAUUCUUC