Counting alignments
This question is about the edit distance problem, which consists in finding the minimum number of edit operations that are
needed to transform one input string into the other. We have seen in class that although there is an exponential number of
alignments, the problem can be efficiently solved in quadratic time and space using dynamic programming. Write a program,
involving a dynamic programming solution, to count the number of possible alignments for two given input sequences of
length m and n respectively.
1. This involves developing a recurrence equation that solves this problem.
2. As well as its implementation.
The input for your program is two numbers, say 280 and 240, which are interpreted as the length of two sequences. The output
is the number of possible sequence alignments for two sequences of such lengths. Note: Make sure that your implementation
works for reasonable sizes of sequences, say 280 and 240.